chore: rename dropdown component
